# improve performance of dynamic sprinkle props