--- a/apps/docs/components/benchmark/index.ts
+++ b/apps/docs/components/benchmark/index.ts
@@ -1,0 +1,1 @@
+export * from './Benchmark';