--- a/apps/docs/pages/styled-system/_meta.json
+++ b/apps/docs/pages/styled-system/_meta.json
@@ -32,6 +32,7 @@
   },
   "state-modifiers": "State Modifiers",
   "theme": "Theme",
+  "benchmarks": "Benchmarks",
   "-- Borders": {
     "type": "separator",
     "title": "Borders"