--- a/apps/docs/pages/styled-system/benchmarks.mdx
+++ b/apps/docs/pages/styled-system/benchmarks.mdx
@@ -1,0 +1,5 @@
+import { Benchmark } from "@/components/benchmark";
+
+# Benchmarks
+
+<Benchmark />