--- a/packages/react/src/css-idents.ts
+++ b/packages/react/src/css-idents.ts
@@ -120,4 +120,6 @@ export default {
   "4o": "outlineOffset",
   "4p": "outlineWidth",
   "4q": "outlineStyle",
+  "4r": "table",
+  "4s": "table-row",
 } as const;