currently we generate inline styles for all sprinkle style props for all media query and state modifier combinations

but we can improve our performance here by pre-generating some of the css while keeping others dynamic - basically striking a balance between the generated css file size vs the performance of most commonly used styles.

so we'll be pre-generating css for the base media query and when no state modifiers are present. for the rest we'll continue to use css var and dynamic style attributes.

## Benchmarks

### Axiom (main) - 26 fps

https://github.com/optimizely-axiom/optiaxiom/assets/758244/f53a016d-fa0e-4736-8760-5c36c82983f2

### CSS Modules (main) - 72 fps

https://github.com/optimizely-axiom/optiaxiom/assets/758244/c5e09a00-60e4-4a74-9b48-2790ea3018df

### styled-jsx (main) - 49 fps

https://github.com/optimizely-axiom/optiaxiom/assets/758244/98c1c5bb-af4d-46c7-b19e-d9de46ab0679

### Axiom (PR) - 60 fps

https://github.com/optimizely-axiom/optiaxiom/assets/758244/8e8152f8-0d45-4bf7-af00-ec18d90eb7d9